renames ipfs-core to just core
