+++ src/cli/utils.js
@@ -2,7 +2,7 @@ const fs = require('fs')
 const os = require('os')
 const APIctl = require('ipfs-api')
 const multiaddr = require('multiaddr')
-const IPFS = require('../ipfs-core')
+const IPFS = require('../core')
 const debug = require('debug')
 const log = debug('cli')
 log.error = debug('cli:error')
