+++ src/core/index.js
