+++ tests/test-core/test-version.js
@@ -3,7 +3,7 @@
 const expect = require('chai').expect
 
 process.env.IPFS_PATH = process.cwd() + '/tests/repo-example'
-const IPFS = require('../../src/ipfs-core')
+const IPFS = require('../../src/core')
 
 describe('version', () => {
   var ipfs
